feat(Cache): multiple performance improvements + deeper implementation
the cache service does now no longer constrain all cache entries to a specific pid, this allows to reuse cached parts on different pages. This might cause issues if elements are page specific, in that case the cache should be disabled for those elements which is acceptable for such edge cases.

The cache service is now present on the most common points of interaction, so it is easy to disable the cache if required.